Contact endpoint, model and migration
It's simple and it works.

There's more I could say such as implementing anti-spam, captcha and so on. But for now, for a simple contact us, it will work. It stores it in a table.

It currently requires no authentication and points to a user. In the future, we might want to make it identify a specific user with an optional user_id column in the table?